# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HFT Bitcoin prediction & paper trading API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/address/{address}:
    get:
      operationId: getWalletAddress
      tags: [wallet]
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
        - name: network
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Address info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddressInfo"
  /wallet/create:
    post:
      operationId: createWallet
      tags: [wallet]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWalletInput"
      responses:
        "200":
          description: Wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletData"
  /wallet/multisig:
    post:
      operationId: createMultisig
      tags: [wallet]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMultisigInput"
      responses:
        "200":
          description: Multisig data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultisigData"
  /wallet/balance:
    post:
      operationId: walletBalance
      tags: [wallet]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletBalanceInput"
      responses:
        "200":
          description: Balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBalance"
  /wallet/send:
    post:
      operationId: walletSend
      tags: [wallet]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletSendInput"
      responses:
        "200":
          description: Send result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletSendResult"
...
  /lightning/decode:
    post:
      operationId: decodeLightningInvoice
      tags: [lightning]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                invoice:
                  type: string
              required: [invoice]
      responses:
        "200":
          description: Decoded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecodedInvoice"
  /lightning/lnurl:
    post:
      operationId: resolveLnurl
      tags: [lightning]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lnurl:
                  type: string
              required: [lnurl]
      responses:
        "200":
          description: LNURL resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LnurlResolved"
  /lightning/lnurl-pay:
    post:
      operationId: lnurlPay
      tags: [lightning]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                callback:
                  type: string
                amountSats:
                  type: number
                comment:
                  type: string
              required: [callback, amountSats]
      responses:
        "200":
          description: LNURL pay result with invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LnurlPayResult"
...
components:
  schemas:
...
    AddressTx:
      type: object
      properties:
        txid: { type: string }
        confirmed: { type: boolean }
        blockHeight:
          oneOf: [{ type: integer }, { type: "null" }]
        blockTime:
          oneOf: [{ type: integer }, { type: "null" }]
        feeSats: { type: number }
        inputs: { type: integer }
        outputs: { type: integer }
      required: [txid, confirmed, feeSats, inputs, outputs]
    AddressInfo:
      type: object
      properties:
        address: { type: string }
        network: { type: string }
        confirmedSats: { type: number }
        unconfirmedSats: { type: number }
        totalSats: { type: number }
        txCount: { type: integer }
        recentTxs:
          type: array
          items: { $ref: "#/components/schemas/AddressTx" }
      required: [address, network, confirmedSats, unconfirmedSats, totalSats, txCount, recentTxs]
    DerivedAddress:
      type: object
      properties:
        index: { type: integer }
        path: { type: string }
        address: { type: string }
        publicKey: { type: string }
        change: { type: integer }
      required: [index, path, address, publicKey, change]
    CreateWalletInput:
      type: object
      properties:
        bip: { type: integer }
        network: { type: string }
        strength: { type: integer }
        passphrase: { type: string }
        mnemonic: { type: string }
        count: { type: integer }
      required: [bip]
    WalletData:
      type: object
      properties:
        bip: { type: integer }
        scheme: { type: string }
        network: { type: string }
        mnemonic: { type: string }
        seedHex: { type: string }
        derivationPath: { type: string }
        xpub: { type: string }
        fingerprint: { type: string }
        receive:
          type: array
          items: { $ref: "#/components/schemas/DerivedAddress" }
        change:
          type: array
          items: { $ref: "#/components/schemas/DerivedAddress" }
      required: [bip, scheme, network, mnemonic, seedHex, derivationPath, xpub, fingerprint, receive, change]
    CreateMultisigInput:
      type: object
      properties:
        m: { type: integer }
        cosigners: { type: integer }
        network: { type: string }
        type: { type: string }
        mnemonics:
          type: array
          items: { type: string }
        passphrases:
          type: array
          items: { type: string }
        publicKeysHex:
          type: array
          items: { type: string }
      required: [m]
    MultisigSigner:
      type: object
      properties:
        index: { type: integer }
        mnemonic: { type: string }
        publicKey: { type: string }
        xpub: { type: string }
        derivationPath: { type: string }
      required: [index, mnemonic, publicKey, xpub, derivationPath]
    MultisigData:
      type: object
      properties:
        m: { type: integer }
        n: { type: integer }
        type: { type: string }
        network: { type: string }
        address: { type: string }
        redeemScript: { type: string }
        witnessScript: { type: string }
        signers:
          type: array
          items: { $ref: "#/components/schemas/MultisigSigner" }
      required: [m, n, type, network, address, redeemScript, signers]
    WalletBalanceInput:
      type: object
      properties:
        mnemonic: { type: string }
        bip: { type: integer }
        network: { type: string }
        scanCount: { type: integer }
        passphrase: { type: string }
      required: [mnemonic, bip]
    WalletBalanceAddress:
      type: object
      properties:
        address: { type: string }
        derivationPath: { type: string }
        sats: { type: number }
        txCount: { type: integer }
      required: [address, derivationPath, sats, txCount]
    WalletBalance:
      type: object
      properties:
        totalSats: { type: number }
        confirmedSats: { type: number }
        unconfirmedSats: { type: number }
        utxoCount: { type: integer }
        addresses:
          type: array
          items: { $ref: "#/components/schemas/WalletBalanceAddress" }
      required: [totalSats, confirmedSats, unconfirmedSats, utxoCount, addresses]
    WalletSendInput:
      type: object
      properties:
        mnemonic: { type: string }
        bip: { type: integer }
        network: { type: string }
        passphrase: { type: string }
        toAddress: { type: string }
        amountSats: { type: number }
        feeRate: { type: number }
        scanCount: { type: integer }
        changeIndex: { type: integer }
      required: [mnemonic, bip, toAddress, amountSats, feeRate]
    WalletSendResult:
      type: object
      properties:
        txid: { type: string }
        hex: { type: string }
        fee: { type: number }
        vsize: { type: number }
        inputsUsed: { type: integer }
        changeSats: { type: number }
      required: [txid, hex, fee, vsize, inputsUsed, changeSats]
...
    DecodedInvoice:
      type: object
      properties:
        network: { type: string }
        paymentHash: { type: string }
        amountSats:
          oneOf: [{ type: number }, { type: "null" }]
        description:
          oneOf: [{ type: string }, { type: "null" }]
        payeeNodeKey:
          oneOf: [{ type: string }, { type: "null" }]
        expiry: { type: number }
        timestamp: { type: number }
        expiresAt: { type: number }
        routeHints: { type: integer }
        features:
          type: array
          items: { type: string }
        raw: { type: string }
      required: [network, paymentHash, expiry, timestamp, expiresAt, routeHints, features, raw]
    LnurlResolved:
      type: object
      properties:
        type: { type: string }
        url: { type: string }
        data:
          type: object
          additionalProperties: true
      required: [type, url, data]
    LnurlPayResult:
      type: object
      properties:
        pr: { type: string }
        routes:
          type: array
          items:
            type: object
            additionalProperties: true
        decoded: { $ref: "#/components/schemas/DecodedInvoice" }
      required: [pr, routes, decoded]
...